*Insert footgun emoji*
https://unicode.org/emoji/proposals.html

Submitting Emoji Proposals

Anyone can submit a proposal for an emoji character, but the proposal
needs to have all the right information for it to have a chance of being
accepted.

This page describes the process of submitting a proposal, including how
to submit a proposal, the selection factors that need to be addressed in
each proposal, guidelines on presenting evidence of frequency, and the
process and timeline for acceptance.

Not all new emoji require new characters. Thus this page also describes
the process of proposing that:

* existing characters be changed to be emoji, or additional emoji
sequences be added as valid or RGI.  Please read this entire page before
preparing a proposal. In particular:

* Scan the list of Emoji Requests to see whether your proposed emoji
have already been submitted. If the status is Declined, it may not be
worth the effort to make a proposal.

* Make sure you understand this whole document, especially the Selection Factors.
Read the Limitations on Emoji Encoding section of Emoji Encoding Principles.

* Read the Emoji Submission FAQ for common questions and answers.

* Review some Example Submissions to see how previous proposals were constructed.

Proposals are likely to be declined unless they are complete and adhere
to the submission instructions. Other proposals may be returned to the
submitter for adjustment based on subcommittee review.